# api/openapi.yaml
...
components:
  schemas:
...
    DeletionRequest:
      description: What an erasure request produced.
      properties:
        eraseAfter:
          description: when the grace period runs out and the erasure begins
          format: date-time
          type: string
        revocationToken:
          description: |-
            the secret that withdraws it. Returned **once**: where
            `plugin-smtp` is installed it also goes out by mail, and where it is not — which a
            `minimal` installation is by design — this is the only way it reaches anybody
          type: string
      type: object
...
    Problem:
      additionalProperties: true
      description: 'An error, as RFC 9457 describes one. A client branches on `type` and never on `detail`:
        the URI is stable and documented under `api/problems/`, the prose is not and is translated (REQ-API-003).'
      properties:
        detail:
          description: What happened this time, in a sentence. Not stable; do not branch on it.
          type: string
        instance:
          description: The path of the request that failed.
          format: uri
          type: string
        status:
          description: The HTTP status, repeated in the body.
          format: int32
        title:
          description: The condition stable, human-readable name.
          type: string
        traceId:
          description: The identifier of this request in the server log. Quoting it in a report leads
            straight to the operation (REQ-NFR-042).
          type: string
        type:
          description: What went wrong, as an identifier. One of the tokens in `docs/reference/problem-types.yaml`,
            under `https://home-inv.example/problems/`. Fixed for the product rather than derived from
            a deployment hostname, so a client recognises it from every instance.
          format: uri
          type: string
      required:
      - status
      - title
      - type
...
    TenantMembershipView:
      description: One of the caller's memberships.
      properties:
        id:
          description: the tenant
          format: uuid
          type: string
        name:
          description: its display name, which is what a switcher shows
          type: string
        role:
          description: the role the caller holds in it
          type: string
      type: object
...
info:
  contact:
    email: lucas.greuloch@gmail.com
    name: Lucas Greuloch (greluc)
    url: https://github.com/greluc/Home-Inventory
  description: |-
    The HTTP surface of a self-hosted Home Inventory instance.

    This document is generated from the implementation and committed (ADR-0049); it describes what the code does rather than what it was meant to do. Errors follow RFC 9457 on every endpoint, and a client branches on the `type` URI, which is registered in `docs/reference/problem-types.yaml` and documented under `api/problems/` (REQ-API-003).

    Paths are versioned. Within a major version a field may be added and an optional one may appear, and nothing a client already reads changes meaning (REQ-API-008).
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
  title: Home Inventory API
  version: v1
openapi: 3.1.0
paths:
...
  /api/v1/me/tenants:
    get:
      description: |-
        Every tenant the caller belongs to, oldest membership first.

        Read through the lookup that works without a tenant context, because the caller is asking
        which tenant to have and therefore cannot be required to already have one.

        There is no cursor, and there does not need to be one: a tenant limit is capped at 200
        (`AppUser.MAX_TENANT_LIMIT`), which is REQ-NFR-010's page size, so one page holds every
        membership anybody can have. The `limit` exists so that a client asking for fewer gets
        fewer, and so that the bound is visible in the contract rather than only in a constraint.
      operationId: tenants
      parameters:
      - description: how many at most; capped at 200, which is also the most anybody can have
        in: query
        name: limit
        required: false
        schema:
          default: 200
          exclusiveMinimum: 0
          format: int32
          maximum: 200
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TenantMembershipView'
                type: array
          description: the memberships, each with the tenant's name and the role held in it
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Malformed request (`malformed-request`)
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthenticated (`unauthenticated`)
        '405':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Method not allowed (`method-not-allowed`)
        '406':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not acceptable (`not-acceptable`)
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation failed (`validation-failed`)
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal error (`internal-error`)
      summary: Every tenant the caller belongs to, oldest membership first
...
  /api/v1/tenants/{tenantId}:
    delete:
      description: |-
        Asks for the tenant to be erased (REQ-TEN-011, REQ-PRIV-005).

        Answered `202`: nothing is erased, and for thirty days nothing will be. The tenant
        stops answering at once — 05 §5.9's "access blocked immediately, data still present" — and the
        answer carries the token that withdraws the request and the instant the erasure begins.

        The token comes back in the response because that is the only way it reaches anybody where
        `plugin-smtp` is not installed, which a `minimal` installation is by design. Where
        it is, the same link goes out by mail.

        `REQ-SEC-021` requires the second factor to be confirmed again for an operation like
        this one. It is wired in with the second factor itself; until then the permission and the
        owner's session are what stands in front of it, and no release ships without the rest.
      operationId: requestDeletion
      parameters:
      - description: the tenant, which must be the session's
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionRequest'
          description: the revocation token and when the erasure begins
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthenticated (`unauthenticated`)
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden (`forbidden`)
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found (`not-found`)
        '405':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Method not allowed (`method-not-allowed`)
        '406':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not acceptable (`not-acceptable`)
        '409':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Deletion pending (`deletion-pending`)
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal error (`internal-error`)
      summary: Asks for the tenant to be erased (REQ-TEN-011, REQ-PRIV-005)
    get:
      description: |-
        What state the tenant is in.

        Reachable while it is blocked, deliberately: a member who is being refused everything else
        is entitled to know why, and 05 §5.9 and open point O26 both say the state is visible to a
        member in the administration view.
      operationId: state
      parameters:
      - description: the tenant, which must be the session's
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantStateView'
          description: the state
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthenticated (`unauthenticated`)
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden (`forbidden`)
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found (`not-found`)
        '405':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Method not allowed (`method-not-allowed`)
        '406':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not acceptable (`not-acceptable`)
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal error (`internal-error`)
      summary: What state the tenant is in
  /api/v1/tenants/{tenantId}/invitations:
    get:
      description: One page of the tenant's invitations, newest first.
      operationId: invitations
      parameters:
      - description: the tenant, which must be the session's
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
      - description: an opaque cursor from a previous page, or omitted for the first
        in: query
        name: cursor
        required: false
        schema:
          maxLength: 500
          minLength: 0
          type: string
      - description: how many at most; capped at 200
        in: query
        name: limit
        required: false
        schema:
          default: 50
          exclusiveMinimum: 0
          format: int32
          maximum: 200
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationPage'
          description: the page, without any tokens
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Malformed request (`malformed-request`)
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthenticated (`unauthenticated`)
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden (`forbidden`)
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found (`not-found`)
        '405':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Method not allowed (`method-not-allowed`)
        '406':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not acceptable (`not-acceptable`)
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation failed (`validation-failed`)
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal error (`internal-error`)
      summary: One page of the tenant's invitations, newest first.
    post:
      description: |-
        Invites somebody into the tenant.

        The answer carries the token, once. Where `plugin-smtp` is installed the invitation
        also goes out by mail; where it is not — which a `minimal` installation is by design —
        this is the only way the link reaches anybody, and 04 §4.3 describes that as the expected
        reduced state rather than a failure.
      operationId: invite
      parameters:
      - description: the tenant, which must be the session's
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteRequest'
        description: the address and the role
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedInvitation'
          description: the invitation and its one token
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Malformed request (`malformed-request`)
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthenticated (`unauthenticated`)
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden (`forbidden`), or Role escalation (`role-escalation`)
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found (`not-found`)
        '405':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Method not allowed (`method-not-allowed`)
        '406':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not acceptable (`not-acceptable`)
        '409':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Invitation already open (`invitation-already-open`)
        '413':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Payload too large (`payload-too-large`)
        '415':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unsupported media type (`unsupported-media-type`)
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation failed (`validation-failed`)
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal error (`internal-error`)
      summary: Invites somebody into the tenant
  /api/v1/tenants/{tenantId}/invitations/{invitationId}:
    delete:
      description: Withdraws an invitation that has not been used.
      operationId: revokeInvitation
      parameters:
      - description: the tenant, which must be the session's
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
      - description: the invitation
        in: path
        name: invitationId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '204':
          description: No Content
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthenticated (`unauthenticated`)
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden (`forbidden`)
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found (`not-found`)
        '405':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Method not allowed (`method-not-allowed`)
        '406':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not acceptable (`not-acceptable`)
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal error (`internal-error`)
      summary: Withdraws an invitation that has not been used.
  /api/v1/tenants/{tenantId}/members:
    get:
      description: One page of the tenant's members.
      operationId: members
      parameters:
      - description: the tenant, which must be the session's
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
      - description: an opaque cursor from a previous page, or omitted for the first
        in: query
        name: cursor
        required: false
        schema:
          maxLength: 500
          minLength: 0
          type: string
      - description: how many at most; capped at 200
        in: query
        name: limit
        required: false
        schema:
          default: 50
          exclusiveMinimum: 0
          format: int32
          maximum: 200
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberPage'
          description: the page
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Malformed request (`malformed-request`)
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthenticated (`unauthenticated`)
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden (`forbidden`)
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found (`not-found`)
        '405':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Method not allowed (`method-not-allowed`)
        '406':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not acceptable (`not-acceptable`)
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation failed (`validation-failed`)
        '500':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Internal error (`internal-error`)
      summary: One page of the tenant's members.
...